testing float and pointer updates

+++ headers/constants.h
@@ -7,7 +7,7 @@
 
 
 constexpr float deg2rad(float _dDeg) {
-    return _dDeg / 180.0f * M_PI;
+    return _dDeg / 180.0f * (float)M_PI;
 }
